requirements.txt
Flask
python-dotenv
requests
flask-sqlalchemy # bridge between Python and SQLite database
alembic # database migration tool